split test hardware code out from AutoLib into its own TestHardware.java file -- update all uses
