+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_Test/_Motors/TestMotorEncoders.java
@@ -5,6 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 import org.firstinspires.ftc.teamcode._Libs.AutoLib;
+import org.firstinspires.ftc.teamcode._Libs.TestHardware;
 
 /**
  * Created by phanau on 2/23/18.
@@ -24,12 +25,12 @@ public class TestMotorEncoders extends OpMode {
 
     public void init() {
         // get hardware
-        AutoLib.HardwareFactory mf = null;
+        TestHardware.HardwareFactory mf = null;
         final boolean debug = false;
         if (debug)
-            mf = new AutoLib.TestHardwareFactory(this);
+            mf = new TestHardware.TestHardwareFactory(this);
         else
-            mf = new AutoLib.RealHardwareFactory(this);
+            mf = new TestHardware.RealHardwareFactory(this);
 
         // get the motors: depending on the factory we created above, these may be
         // either dummy motors that just log data or real ones that drive the hardware
