+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_Test/_Motors/TestServoSteps.java
@@ -6,6 +6,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.teamcode._Libs.AutoLib;
+import org.firstinspires.ftc.teamcode._Libs.TestHardware;
 
 /**
  * Created by phanau on 12/7/18.
@@ -32,12 +33,12 @@ public class TestServoSteps extends OpMode {
 
     public void init() {
         // get hardware
-        AutoLib.HardwareFactory mf = null;
+        TestHardware.HardwareFactory mf = null;
         final boolean debug = false;
         if (debug)
-            mf = new AutoLib.TestHardwareFactory(this);
+            mf = new TestHardware.TestHardwareFactory(this);
         else
-            mf = new AutoLib.RealHardwareFactory(this);
+            mf = new TestHardware.RealHardwareFactory(this);
 
         // get the servo: depending on the factory we created above, these may be
         // either dummy motors that just log data or real ones that drive the hardware
