+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_Test/_Sensors/DeadWheelTest.java
@@ -4,6 +4,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.teamcode._Libs.AutoLib;
+import org.firstinspires.ftc.teamcode._Libs.TestHardware;
 
 
 /**
@@ -21,7 +22,7 @@ public class DeadWheelTest extends OpMode {
 
     public void init() {
         // get hardware
-        AutoLib.HardwareFactory mf = new AutoLib.RealHardwareFactory(this);
+        TestHardware.HardwareFactory mf = new TestHardware.RealHardwareFactory(this);
         mMotor = mf.getDcMotor("bl");       // motor 0 in reverseRatbot configuration
     }
